todo_tasks/wider_todo/Version 1/requirements.txt
django==3.1.5
django-crontab
python-dateutil
